+++ changelog.md
@@ -4,6 +4,7 @@
 - added a quartz -> refined quartz smelting and blasting recipe
 - added a custom advancements tab for chapter 1
 - removed minecraft/story advancements
+- removed experimental features screen
 
 ## [0.0.19]
 ### Changes